+++ test/novident_nodes_test.dart
@@ -100,8 +100,7 @@ void main() {
     if (Node.canMoveTo(
       node: node.first,
       target: node.last,
-      inside: true,
-      isSwapMove: true,
+      position: DropPosition.inside,
     )) {
       moved = Node.moveTo(
           node: node.first, newOwner: node.last.castToContainer, index: 1);